requirements.txt
cycler==0.11.0
fonttools==4.28.5
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.21.5
opencv-python==4.5.4.60
packaging==21.3
pandas==1.3.5
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
split-folders==0.4.3
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
typing_extensions==4.0.1
